Fix logic error causing first card not to have a sprite in the spawned deck.
Issue was that the sprite slices from the card faces spritesheet are not
loaded in order. Because the original loop was confined to 52 and not
the entire 54 (e.g., length of the spritesheet array) the sprite for the
first card was inadvertently excluded.

Next step is to shuffle the deck