Fix bug #2. Add constraint check to the function that gets the next available move to not allow more than one dragged card to be transitioned to the foundations in a single event.
Next steps are to enhance the translate animation system to make all
translation take a set amount of time so that the speed is relative to
the total distance that needs to be traversed against the total time it
should take to get there. For example, if the total amount of time for
all translation animations should be 1 second then the translation speed
to a card to go from point A to point B where the distance between point
A and point B is large will result in a faster translation speed than if
the distance between point A and point B is small.