Fix animation issues with timer label show/hide All animations are fixed for portrait and landscape orientations.
Next steps are to address issue with game states being lost between
orientations. For example, if the timer label is hidden in portrait, it
is visible again in landscape. Also, a more serious issue is that you
can open a pause menu in one orientation and it won't be open in the
other orientation, causing a lock.